Added in automated getter and setter testing into Middleware
This was done so that any new pojos added are automatically tested. This was inspired so that we cover the XADataSourceProperties.

Fixed some pojo's which were calling equals on their super class which was the Object class.
Added in missing setters if required
Fixed null pointer exceptions as a result of the tests added.

issue: BMS-1369
reviewer: MatthewB