Updating schema to enable more performant inserts by dropping the need to insert duplicated data into the unique name column
We have made unique name nullable and dropped the unique index on it. The eventual plan is to get rid of the column.

BMS will no longer populate this column as the unique name is populated with exactly the same data as the phenotype_id. Comment below from Graham on why this is the case.

Comments from Graham.

Indeed you are correct, there is some information duplicated in the phenotype table. This was just a problem of fitting our data model into the Chado schema without changing it. Some fields are compulsory and we had no matching information to store there so we repeated some fields. One interesting consequence is that we have some information stored both as text and number and this can be useful because the cvterm_id stored in the observable_id field sometimes needs to link to the cvterm_id stored in the projectprop.value field when it is cast as text.

Also added in a sequence table so that hibernated can generate ID's in memory.

issue: BMS-1797
reviewer: MatthewB