Fixed Incorrect Variable Usage Count
Saving of trials or nurseries let to adding of an additional row in the phenotype table. This was because after adding a phenotype record we did not update the phenotype id in the UI. This is now achieved by setting in the phenotype id into each measurement data on creation or updating.

issue: BMS-1771
reviewer: MatthewB