Fixed saving issues related to the BM_CODE_VTE selection variate
We can now store breeding methods associated with each plot. This was causing an error because the value was being stored as a categorical variable while the BM_CODE_VTE is declared as a character variable in the UI. Have added in additional checks when saving a phenotype value.

issue: BMS-2361
reviewer: MatthewB