Updating gemplasm entity using saveOrUpdate method and adding new test cases
Calling save() on germplasm entity/object results in new germplasm entity(with new gid) because germplasm entity  is saved in different hibernate session and so act as non-persitenet. This results in hibernate constraint violation error. After pagination loads germplasms, the hibernate session in which it is loaded is not the one in which we save/update germplasm.

Calling saveOrUpdate() on germplasm behave correctly as that will work in both persistent and non-persistent entity state. Irrespctive of hibernate session germplasms are loaded and saved correctly.

issue: BMS-2724
reviewer: Sunny