Removing heavy variable observation calculation
Variable observation calculation is heavy operation as it calculates based on different variable types from different table and setting it to 0. To define if varible is editable/deletable, observation count is not needed. If we get variable study count(from projectprop) then we can clearly identify if variable is used or not and determine editable/deletable or not.

issue: BMS-3107
reviewer: Sunny