Introduce relation between users and crops
In relation to the feature of restricting users to work on particular
crops, part of the Role and Permissions epic.
WorkbenchUser is an entity widely used across BMS, sometimes even
outside the scope of a transaction. Because of this, the mapping has
EAGER fetching. FetchMode.SUBSELECT is also needed because it's a List.
Display the list of crops in the users table in Site admin.

Issue: IBP-2697
Reviewer: None