Lot label printing: first draft
LotLabelPrinting does not consider TermId for now. It uses a fixed set
of columns and enumerates the rest using an offset.
E.g.
   LOT_UUID: id=1
   ...
   CROSS: id=21
   ATTRIBUTE1: id=(udflds.fldno + 21)

This works for now as the only variable columns are germpl attributes.
To support a different type of variable column another extension
will need to be devised.

searchRequestId is added as parameter because it could be use with other
types of label export.

Issue: IBP-3514