Fix order by performance. Limit results if sorting
If sort is applied, switch to the old-style search-like behaviour, where
the query is limited by a maximum amount of records and those are then
sorted. In some cases, depending on the query optimizer and the usage of
indices, the sort could still take longer than other operations, like
filtering.

Issue: IBP-4021