Add common workaround to pedigree service
To get around the 5000 problem in crossExpansions/Pedigree strings a
workaround has been adopted and used frequently in the application,
which is to used invoke the method several times with lists of 5000
gids. This method is to just have the workaround in one place, while we
continue working on better options.

Issue: IBP-4114