Fix pedigree retrieval when gpid1 or gpid2 is 0
Some breeding scenarios might produce these records in the database, or
they might be the result of data loading errors, but in any case, it's
good to include them in the results.
The restriction (gpid1 !=0, gpid2 != 0) was there because in some older
version of the query, when pedigree was a subquery, there was a
significant improvement when adding these clauses. In the new version,
that receives the list of filtered gids, this is no longer the case.

Issue: IBP-4114