Remove obsolete index germplsm_idx13 and changeset
It has been detected that germplsm_idx13 affected the performance of
germplasm search filtered count in the uat server.
This index was introduced mainly for the total count, but we saw
omitting this count will greatly improve the overall performance of the
search, and since this count is not of much value for the user, we
remove it altogether.

Middleware
- Remove changeset so that we avoid running the expensive (now
obsolete) index creation for each migration.
- Remove created index wherever the old changeset has run.
BMSAPI / Workbench:
- Remove total count.

Issue: IBP-4241