IBP-5352: Fix count No of generations for incomplete derivative pedigree
Parents were not being retrieved for a derivative line with incomplete
records (gnpgs=-1, gpid2=0, gpid1=<groupSourceGid>) and full pedigree
parameter (level=null) set, because the correct level was not calculated
This should not be a problem given that a similar logic (bypass male
parent if unknown) is being performed in addParents.

This is needed for COP calculation that call for full pedigree even for
derivative lines.

TODO test