IBP-5483: BTYPE=2 Self Fertilizing with unknown source but known parents
This case is not described in the paper but was explained in a graham's
email, and it also appears in BROWSE code.
The F4 assumption for break in the records should only be considered for
 btype=2.
From Graham: "It results from a modification to the boundary conditions
when you need a COI for a line which has unknown source but known
parents - so you dont know the number of inbreeding generations.
When BTYPE = 1 this should just be treated as fully inbred, but if BTYPE
is set to 2 then the assumption is that it is a breeding line which are
often exchanged at F4 rather than fully inbred. Only one breeder asked
for this.."

FIXME BTYPE=2 produces COP > 1